# Backend/app/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Consumewise Product API
  description: API for extracting product details and generating personalized health suggestions using Ollama and MongoDB.
  version: 1.0.0

servers:
  - url: http://localhost:5000
    description: Local development server

paths:
  /products:
    get:
      summary: Get all products
      description: Retrieve all stored products from the database.
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      summary: Add a new product
      description: Insert a new product into the database if the barcode is not already present.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "201":
          description: Product added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: Product already exists or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /gethealthsuggestion:
    post:
      summary: Get health suggestions for a product
      description: |
        Generate health suggestions based on either product details from the database (via barcode) 
        or an uploaded product image.  
        Considers nutritional information, ingredients, and user health concerns.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Optional product image file
                data:
                  type: object
                  properties:
                    barcode_number:
                      type: string
                      example: "8901234567890"
                    diseases:
                      type: array
                      items:
                        type: string
                      example: ["diabetes", "hypertension"]
      responses:
        "200":
          description: Health advice generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  advice:
                    type: string
        "400":
          description: Product not found and no image provided or invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /extract_product_details:
    post:
      summary: Extract product details from images
      description: |
        Upload two product images and extract structured product information using Ollama.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image1
                - image2
              properties:
                image1:
                  type: string
                  format: binary
                  description: First product image
                image2:
                  type: string
                  format: binary
                  description: Second product image
      responses:
        "200":
          description: Product details extracted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductWithHealth"
        "400":
          description: Missing images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    Product:
      type: object
      properties:
        item_name:
          type: string
          example: "Oatmeal Cookies"
        brand:
          type: string
          example: "Healthy Bites"
        barcode_number:
          type: string
          example: "8901234567890"
        weight:
          type: string
          example: "200g"
        ingredients:
          type: array
          items:
            type: string
          example: ["Oats", "Sugar", "Butter", "Raisins"]
        nutritional_info:
          type: object
          additionalProperties:
            type: string
          example:
            calories: "120 kcal"
            protein: "5g"
            fat: "2g"
        product_description:
          type: string
          example: "Crispy and healthy oatmeal cookies with raisins."
    ProductWithHealth:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            health_suggestion:
              type: string
              example: "Good source of fiber but avoid if diabetic due to sugar content"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Product with this barcode already exists!"